Update detect_volumes_to_adjust.py for performance
Added a cache that avoids useless repetitions of volume calculations (expensive). Before, when a component had more than one body, the volume of all the bodies would be calculated as many times as the number of bodies. Now it will be calculated only once.

Also, if a component has the same name and hierarchy tree as other, it is assumed they are identical, its volume won't be calculated.

This change reduced the amount of time to run the script in a certain CAD package from almost an hour to aprox. 1 minute.